Feat supporting Token distribution and verification
Fixed: #20

This commit first categorizes the KBS HTTP API into three categories and reorganizes the code accordingly:
1. attest: RESTful APIs that related to attestation
2. resource: RESTful APIs that to get secret resources, need attestation verification
3. config: RESTful APIs that configure KBS and AS, require user authentication
4. public: RESTful APIs that is public

This commit then implement the Token distribution and verification function module.
