+++ config/kbs-config.json
@@ -3,5 +3,6 @@
     "repository_description": {
         "dir_path": "/opt/confidential-containers/kbs/repository"
     },
+    "attestation_token_type": "Simple",
     "as_addr": "http://as:50004"
 }
