+++ src/api/Cargo.toml
@@ -25,6 +25,7 @@ as-types = { git = "https://github.com/confidential-containers/attestation-servi
 attestation-service = { git = "https://github.com/confidential-containers/attestation-service.git", default-features = false, rev = "v0.5.0", optional = true}
 base64.workspace = true
 cfg-if = "1.0.0"
+elliptic-curve = { version = "0.13.4", features = ["arithmetic", "pem"] }
 env_logger.workspace = true
 jwt-simple = "0.11.4"
 kbs-types = "0.2"
@@ -32,6 +33,7 @@ lazy_static = "1.4.0"
 log.workspace = true
 prost = { version = "0.11", optional = true }
 rand = "0.8.5"
+rcgen = "0.10.0"
 rsa = "0.7.2"
 rustls = { version = "0.20.8", optional = true }
 rustls-pemfile = { version = "1.0.2", optional = true }
@@ -44,6 +46,7 @@ tokio.workspace = true
 tonic = { version = "0.8", optional = true }
 uuid = { version = "1.2.2", features = ["serde", "v4"] }
 openssl = { version = "0.10.46", optional = true }
+p256 = { version = "0.13.2", features = [ "ecdsa" ] }
 
 [dev-dependencies]
 tempfile.workspace = true
