+++ src/api/src/resource/mod.rs
@@ -2,17 +2,10 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::rand::Rng;
-use aes_gcm::{
-    aead::{Aead, KeyInit, OsRng},
-    Aes256Gcm, Nonce,
-};
-use anyhow::{anyhow, Result};
-use kbs_types::{Response, TeePubKey};
+use anyhow::*;
 use local_fs::{LocalFs, LocalFsRepoDesc};
-use rsa::{BigUint, PaddingScheme, PublicKey, RsaPublicKey};
 use serde::Deserialize;
-use serde_json::{json, Value};
+use serde_json::Value;
 use std::fs;
 use std::path::Path;
 use std::sync::Arc;
@@ -20,9 +13,6 @@ use strum_macros::EnumString;
 use tokio::sync::RwLock;
 
 mod local_fs;
-
-const RSA_ALGORITHM: &str = "RSA1_5";
-const AES_GCM_256_ALGORITHM: &str = "A256GCM";
 
 /// Interface of a `Repository`.
 #[async_trait::async_trait]
@@ -76,60 +66,6 @@ impl RepositoryType {
             }
         }
     }
-}
-
-pub(crate) async fn get_secret_resource(
-    tee_pub_key: TeePubKey,
-    repository: &Arc<RwLock<dyn Repository + Send + Sync>>,
-    resource_desc: ResourceDesc,
-) -> Result<Response> {
-    if tee_pub_key.alg != *RSA_ALGORITHM {
-        return Err(anyhow!("Unsupported TEE Pub Key type or algorithm"));
-    }
-
-    let resource_byte = repository
-        .read()
-        .await
-        .read_secret_resource(resource_desc)
-        .await
-        .map_err(|e| anyhow!("Read secret resource from repository failed: {:?}", e))?;
-
-    let mut rng = rand::thread_rng();
-
-    let aes_sym_key = Aes256Gcm::generate_key(&mut OsRng);
-    let cipher = Aes256Gcm::new(&aes_sym_key);
-    let iv = rng.gen::<[u8; 12]>();
-    let nonce = Nonce::from_slice(&iv);
-    let encrypted_resource_payload = cipher
-        .encrypt(nonce, resource_byte.as_slice())
-        .map_err(|e| anyhow!("AES encrypt Resource payload failed: {:?}", e))?;
-
-    let k_mod = base64::decode(&tee_pub_key.k_mod)?;
-    let n = BigUint::from_bytes_be(&k_mod);
-    let k_exp = base64::decode(&tee_pub_key.k_exp)?;
-    let e = BigUint::from_bytes_be(&k_exp);
-
-    let rsa_pub_key = RsaPublicKey::new(n, e)
-        .map_err(|e| anyhow!("Building RSA key from modulus and exponent failed: {:?}", e))?;
-    let padding = PaddingScheme::new_pkcs1v15_encrypt();
-    let sym_key: &[u8] = aes_sym_key.as_slice();
-    let wrapped_sym_key = rsa_pub_key
-        .encrypt(&mut rng, padding, sym_key)
-        .map_err(|e| anyhow!("RSA encrypt sym key failed: {:?}", e))?;
-
-    let protected_header = json!(
-    {
-       "alg": RSA_ALGORITHM.to_string(),
-       "enc": AES_GCM_256_ALGORITHM.to_string(),
-    });
-
-    Ok(Response {
-        protected: serde_json::to_string(&protected_header)?,
-        encrypted_key: base64::encode_config(wrapped_sym_key, base64::URL_SAFE_NO_PAD),
-        iv: base64::encode_config(iv, base64::URL_SAFE_NO_PAD),
-        ciphertext: base64::encode_config(encrypted_resource_payload, base64::URL_SAFE_NO_PAD),
-        tag: "".to_string(),
-    })
 }
 
 pub(crate) async fn set_secret_resource(
